+++ backend/chat/model.py
@@ -1,5 +1,46 @@
-import tensorflow as tf 
+import tensorflow as tf
 import numpy as np
+import pickle
+
+# Just disables the warning, doesn't enable AVX/FMA
+import os
+os.environ['TF_CPP_MIN_LOG_LEVEL'] = '2'
+
+# Load in data structures
+with open("chat/data/wordList.txt", "rb") as fp:
+    wordList = pickle.load(fp)
+wordList.append('<pad>')
+wordList.append('<EOS>')
+
+# Load in hyperparamters
+vocabSize = len(wordList)
+batchSize = 24
+maxEncoderLength = 15
+maxDecoderLength = 15
+lstmUnits = 112  # was 48 but should be 112 to match Seq2Seq.py
+numLayersLSTM = 3
+
+# Create placeholders
+encoderInputs = [tf.placeholder(tf.int32, shape=(None,)) for i in range(maxEncoderLength)]
+decoderLabels = [tf.placeholder(tf.int32, shape=(None,)) for i in range(maxDecoderLength)]
+decoderInputs = [tf.placeholder(tf.int32, shape=(None,)) for i in range(maxDecoderLength)]
+feedPrevious = tf.placeholder(tf.bool)
+
+encoderLSTM = tf.nn.rnn_cell.BasicLSTMCell(lstmUnits, state_is_tuple=True)
+#encoderLSTM = tf.nn.rnn_cell.MultiRNNCell([singleCell]*numLayersLSTM, state_is_tuple=True)
+decoderOutputs, decoderFinalState = tf.contrib.legacy_seq2seq.embedding_rnn_seq2seq(encoderInputs, decoderInputs, encoderLSTM,
+                                                            vocabSize, vocabSize, lstmUnits, feed_previous=feedPrevious)
+
+decoderPrediction = tf.argmax(decoderOutputs, 2)
+
+# Start session and get graph
+sess = tf.Session()
+#y, variables = model.getModel(encoderInputs, decoderLabels, decoderInputs, feedPrevious)
+
+# Load in pretrained model
+saver = tf.train.Saver()
+saver.restore(sess, tf.train.latest_checkpoint('chat/models'))
+zeroVector = np.zeros((1), dtype='int32')
 
 def getTestInput(inputMessage, wList, maxLen):
 	encoderMessage = np.full((maxLen), wList.index('<pad>'), dtype='int32')
@@ -35,3 +76,16 @@ def idsToSentence(ids, wList):
     chosenString = listOfResponses[0]
     #chosenString = max(listOfResponses, key=len)
     return chosenString
+
+def pred(inputString):
+    inputVector = getTestInput(inputString, wordList, maxEncoderLength)
+    feedDict = {encoderInputs[t]: inputVector[t] for t in range(maxEncoderLength)}
+    feedDict.update({decoderLabels[t]: zeroVector for t in range(maxDecoderLength)})
+    feedDict.update({decoderInputs[t]: zeroVector for t in range(maxDecoderLength)})
+    feedDict.update({feedPrevious: True})
+    ids = (sess.run(decoderPrediction, feed_dict=feedDict))
+    return idsToSentence(ids, wordList)
+
+def prediction(msg):
+    response =  pred(msg)
+    return response
